requirements.txt
streamlit
mediapipe==0.10.11
opencv-python-headless
gspread
oauth2client
pandas
numpy
Pillow
streamlit-webrtc
av